METRON-1621: moving AutomationHelper to e2e_utils
